requirements.txt
requests==2.31.0
pandas==2.1.0
pyarrow==14.0.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-mock==3.11.1
duckdb==0.9.1
pandera==0.17.2
